# prometheus: Refactor getAttrs